# packages/hotel-pms-simulation/openapi.source.yaml
...
openapi: 3.0.3
info:
  title: Hotel PMS API
  description: 'Hotel Property Management System API for managing hotel reservations and guest services.


    Use this API to help guests find available rooms, get pricing quotes, make reservations,

    and manage their hotel stays. The API supports the complete guest journey from initial

    availability inquiry through checkout.

    '
  version: 1.0.0
  contact:
    name: Hotel PMS API Support
components:
  schemas:
    AvailabilityRequest:
      type: object
      required:
      - hotel_id
      - check_in_date
      - check_out_date
      - guests
      properties:
        hotel_id:
          type: string
          description: Unique identifier for the hotel
          example: H-PVR-002
        check_in_date:
          type: string
          format: date
          description: Check-in date in YYYY-MM-DD format
          example: '2024-03-15'
        check_out_date:
          type: string
          format: date
          description: Check-out date in YYYY-MM-DD format
          example: '2024-03-17'
        guests:
          type: integer
          minimum: 1
          maximum: 10
          description: Number of guests
          example: 2
        package_type:
          type: string
          enum:
          - simple
          - detailed
          description: Response detail level (simple or detailed)
          default: simple
    AvailabilityResponse:
      type: object
      properties:
        hotel_id:
          type: string
          description: Hotel identifier
        check_in_date:
          type: string
          format: date
          description: Check-in date
        check_out_date:
          type: string
          format: date
          description: Check-out date
        guests:
          type: integer
          description: Number of guests
        available_room_types:
          type: array
          description: List of available room types with pricing
          items:
            type: object
            properties:
              room_type_id:
                type: string
                description: Room type identifier
              available_rooms:
                type: integer
                description: Number of available rooms
              base_rate:
                type: number
                format: float
                description: Base rate per night
              total_cost:
                type: number
                format: float
                description: Total cost for the stay (only in detailed package)
    QuoteRequest:
      type: object
      required:
      - hotel_id
      - room_type_id
      - check_in_date
      - check_out_date
      - guests
      properties:
        hotel_id:
          type: string
          description: Unique identifier for the hotel
          example: H-PVR-002
        room_type_id:
          type: string
          description: Unique identifier for the room type
          example: RT-STD
        check_in_date:
          type: string
          format: date
          description: Check-in date in YYYY-MM-DD format
          example: '2024-03-15'
        check_out_date:
          type: string
          format: date
          description: Check-out date in YYYY-MM-DD format
          example: '2024-03-17'
        guests:
          type: integer
          minimum: 1
          maximum: 10
          description: Number of guests
          example: 2
        package_type:
          type: string
          enum:
          - simple
          - detailed
          description: Response detail level (simple or detailed)
          default: simple
...
paths:
  /availability/check:
    post:
      summary: Check room availability
      description: 'Find out what rooms are available at a hotel for specific dates and number of guests.

        Use this when a guest wants to see their options before booking. You''ll get back

        a list of room types that have availability, along with how many rooms are free

        and the nightly rate for each type. This is always the first step before making

        a reservation.

        '
      operationId: check_availability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityRequest'
            examples:
              simple_check:
                summary: Simple availability check
                value:
                  hotel_id: H-PVR-002
                  check_in_date: '2024-03-15'
                  check_out_date: '2024-03-17'
                  guests: 2
              detailed_check:
                summary: Detailed availability check with total costs
                value:
                  hotel_id: H-PVR-002
                  check_in_date: '2024-03-15'
                  check_out_date: '2024-03-17'
                  guests: 2
                  package_type: detailed
      responses:
        '200':
          description: Availability check completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
              examples:
                available_rooms:
                  summary: Available rooms found
                  value:
                    hotel_id: H-PVR-002
                    check_in_date: '2024-03-15'
                    check_out_date: '2024-03-17'
                    guests: 2
                    available_room_types:
                    - room_type_id: RT-STD
                      available_rooms: 5
                      base_rate: 150.0
                    - room_type_id: RT-DLX
                      available_rooms: 2
                      base_rate: 200.0
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /quotes/generate:
    post:
      summary: Generate pricing quote
      description: 'Get an exact price quote for a specific room type and dates. Use this when

        a guest has decided on a room type and wants to see the total cost breakdown.

        The quote gives you a quote_id that you must use to create the actual reservation.

        Quotes are valid for 6 hours, so the guest needs to book within that time.

        '
      operationId: generate_quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
            examples:
              simple_quote:
                summary: Simple quote generation
                value:
                  hotel_id: H-PVR-002
                  room_type_id: RT-STD
                  check_in_date: '2024-03-15'
                  check_out_date: '2024-03-17'
                  guests: 2
              detailed_quote:
                summary: Detailed quote with pricing breakdown
                value:
                  hotel_id: H-PVR-002
                  room_type_id: RT-STD
                  check_in_date: '2024-03-15'
                  check_out_date: '2024-03-17'
                  guests: 2
                  package_type: detailed
      responses:
        '200':
          description: Quote generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
              examples:
                generated_quote:
                  summary: Successfully generated quote
                  value:
                    hotel_id: H-PVR-002
                    room_type_id: RT-STD
                    check_in_date: '2024-03-15'
                    check_out_date: '2024-03-17'
                    guests: 2
                    nights: 2
                    base_rate: 150.0
                    total_cost: 300.0
                    quote_id: Q-20240315-ABC123
                    expires_at: '2024-03-15T18:00:00Z'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservations:
    post:
      summary: Create reservation
      description: 'Book a hotel room for a guest using a quote_id from a previous quote.

        This finalizes the reservation and confirms the booking. You need the

        guest''s name, email, and optionally their phone number. The quote must

        still be valid (not expired) to complete the booking.

        '
      operationId: create_reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
            examples:
              new_reservation:
                summary: Create new reservation
                value:
                  quote_id: Q-20240315-ABC123
                  guest_name: John Doe
                  guest_email: john.doe@example.com
                  guest_phone: +1-555-123-4567
      responses:
        '200':
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
              examples:
                created_reservation:
                  summary: Successfully created reservation
                  value:
                    reservation_id: R-20240315-XYZ789
                    status: confirmed
                    hotel_id: H-PVR-002
                    guest_name: John Doe
                    guest_email: john.doe@example.com
                    guest_phone: +1-555-123-4567
                    check_in_date: '2024-03-15'
                    check_out_date: '2024-03-17'
                    room_type_id: RT-STD
                    total_cost: 300.0
                    created_at: '2024-03-15T12:00:00Z'
        '400':
          description: Invalid request parameters or expired quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get reservations
      description: 'Find existing reservations by searching for a guest''s email or looking up

        all reservations at a specific hotel. Use this when a guest asks about

        their booking or when hotel staff need to see current reservations.

        You can limit how many results you get back.

        '
      operationId: get_reservations
      parameters:
      - name: hotel_id
        in: query
        schema:
          type: string
        description: Filter reservations by hotel ID
        example: H-PVR-002
      - name: guest_email
        in: query
        schema:
          type: string
          format: email
        description: Filter reservations by guest email
        example: john.doe@example.com
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
        description: Maximum number of reservations to return
        example: 10
      responses:
        '200':
          description: Reservations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  reservations:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReservationResponse'
                  total_count:
                    type: integer
                    description: Total number of matching reservations
              examples:
                reservations_list:
                  summary: List of reservations
                  value:
                    reservations:
                    - reservation_id: R-20240315-XYZ789
                      status: confirmed
                      hotel_id: H-PVR-002
                      guest_name: John Doe
                      guest_email: john.doe@example.com
                      check_in_date: '2024-03-15'
                      check_out_date: '2024-03-17'
                      room_type_id: RT-STD
                      total_cost: 300.0
                    total_count: 1
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /reservations/{reservation_id}/checkout:
    post:
      summary: Checkout guest
      description: 'Complete a guest''s checkout when they''re leaving the hotel. Use this

        to finalize their stay and calculate the final bill. You can add any

        extra charges (like room service or minibar) and specify how they''re

        paying (card, cash, or bank transfer).

        '
      operationId: checkout_guest
      parameters:
      - name: reservation_id
        in: path
        required: true
        schema:
          type: string
        description: Unique reservation identifier
        example: R-20240315-XYZ789
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
            examples:
              simple_checkout:
                summary: Simple checkout without additional charges
                value:
                  payment_method: card
              checkout_with_charges:
                summary: Checkout with additional charges
                value:
                  additional_charges: 75.5
                  payment_method: card
      responses:
        '200':
          description: Checkout completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
              examples:
                checkout_complete:
                  summary: Successful checkout
                  value:
                    reservation_id: R-20240315-XYZ789
                    status: checked_out
                    final_bill: 375.5
                    checkout_time: '2024-03-17T11:00:00Z'
        '404':
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid checkout parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...